BACKPORT: media: v4l: fix uapi mpeg slice params definition
We get a headers_check warning about the newly defined ioctl command
structures:

./usr/include/linux/v4l2-controls.h:1105: found __[us]{8,16,32,64} type without #include <linux/types.h>

This is resolved by including linux/types.h, as suggested by the
warning, but there is another problem: Three of the four structures
have an odd number of __u8 headers, but are aligned to 32 bit in the
v4l2_ctrl_mpeg2_slice_params, so we get an implicit padding byte
for each one. To solve that, let's add explicit padding that can
be set to zero and verified in the kernel.

Fixes: c27bb30e7b6d ("media: v4l: Add definitions for MPEG-2 slice format and metadata")

Signed-off-by: Arnd Bergmann <arnd@arndb.de>
Signed-off-by: Mauro Carvalho Chehab <mchehab+samsung@kernel.org>
(cherry picked from commit df18bfd35bbf7cb1a420b5beede1de29343793b3)

Conflicts:
   drivers/media/v4l2-core/v4l2-ctrls.c
   include/uapi/linux/v4l2-controls.h
The only thing actually applicable from the upstream fix was the
move of the #include.  Specifically:
- That patch added padding to 3 structures, but if you compare the
  ones in our tree to mainline Linux (which includes the relevant fix)
  the structures match.  That's because the padding was removed by
  commit d998e03e322f ("media: cedrus: identify buffers by
  timestamp").
- That patch added code to check the 'pad'.  That code was removed by
  the same patch which removed the 'pad' (obviously, since it wouldn't
  compile otherwise).

BUG=chromium:999585
TEST=Build and boot

Change-Id: I6c3949419fd2a64b12e2f158a1a54e82160dddfe
Signed-off-by: Douglas Anderson <dianders@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/1816996
Reviewed-by: Guenter Roeck <groeck@chromium.org>
Reviewed-by: Tomasz Figa <tfiga@chromium.org>
Commit-Queue: Tomasz Figa <tfiga@chromium.org>