CHROMIUM: iwl7000: pcie: remove unnecessary check for null pointer
We only call the iwl_pcie_apply_destination() after checking if
iwl_pcie_dbg_on() returns true.  This function will only return true
if either ini_valid is set or if dbg_dest_tlv is set.  At the
beginning of the function, we check if ini_valid is set, do some stuff
and return.  So if we continue in this function, dbg_dest_tlv must be
not NULL and the check in the for loop is useless.  We even assume
it's non-NULL before the for loop, so there's no point in cheking for
it there.

Change-Id: I4fd9c3912c32df10146ddb056102569513c93873
Signed-off-by: Luca Coelho <luciano.coelho@intel.com>
iwl7000-tree: 0bb9f305489cd348a239e9ef607f1ccafa627324