FROMGIT: arm64: dts: qcom: sdm845-mtp: Mark protected gcc clocks
As of v4.20-rc1 probing the GCC driver on a SDM845 device with the
standard security implementation causes an access violation and an
immediate system restart. Use the protected-clocks property to mark the
offending clocks protected for the MTP, in order to allow it to boot.

Cc: Stephen Boyd <sboyd@kernel.org>
Signed-off-by: Bjorn Andersson <bjorn.andersson@linaro.org>
Acked-by: Andy Gross <andy.gross@linaro.org>
Signed-off-by: Stephen Boyd <sboyd@kernel.org>
(cherry picked from commit d206e6b7ea3fdc8ec8f6be9a2ecfe58142b49e37
 git://git.kernel.org/pub/scm/linux/kernel/git/clk/linux.git clk-next)

BUG=b:112880092
TEST=Build and boot

Change-Id: Ib44b60f5190f763e6eb9b5db9f73b147ee396033
Signed-off-by: Douglas Anderson <dianders@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1367033
Reviewed-by: Stephen Boyd <swboyd@chromium.org>