FROMLIST: wilco_ec: Add new driver for Wilco EC
(am from https://lkml.org/lkml/2019/1/23/672)

This EC is an incompatible variant of the typical Chrome OS embedded
controller.  It uses the same low-level communication and a similar
protocol with some significant differences.  The EC firmware does
not support the same mailbox commands so it is not registered as a
cros_ec device type.

BUG=b:119046283
TEST=manual

Change-Id: If32b29c93bcf94bff976d23b686c20b8433a46c1
Signed-off-by: Duncan Laurie <dlaurie@google.com>
Acked-for-chrome-platform-by: Enric Balletbo i Serra <enric.balletbo@collabora.com>
Reviewed-on: https://chromium-review.googlesource.com/1323830
Commit-Ready: Nick Crews <ncrews@chromium.org>
Tested-by: Nick Crews <ncrews@chromium.org>
Reviewed-by: Nick Crews <ncrews@chromium.org>