CHROMIUM: iwl7000: mac80211: fix radiotap vendor presence bitmap handling
Due to the alignment handling, it actually matters where in the code
we add the 4 bytes for the presence bitmap to the length; the first
field is the timestamp with 8 byte alignment so we need to add the
space for the extra vendor namespace presence bitmap *before* we do
any alignment for the fields.

Move the presence bitmap length accounting to the right place to fix
the alignment for the data properly.

Change-Id: Ie8e344dd2a720665a28b76f02ea988ed3beb2bf2
Signed-off-by: Johannes Berg <johannes.berg@intel.com>
iwl7000-tree: 77c617f5facc15716e9ff6151a8b77e104a19f5a