FROMGIT: drm/msm: dpu: Move crtc runtime resume to encoder
The crtc runtime resume doesn't actually operate on the crtc, but rather
its encoders. The problem with this is that we need to inspect the crtc
state to get the currently connected encoders. Since runtime resume
isn't guaranteed to be called while holding the modeset locks (although
it sometimes is), this presents a race condition.

Now that we have ->enabled on the virtual encoders, and a lock to
protect it, just call resume on each encoder and only restore the ones
that are enabled.

Changes in v2:
- None

Cc: Jeykumar Sankaran <jsanka@codeaurora.org>
Reviewed-by: Jeykumar Sankaran <jsanka@codeaurora.org>
Signed-off-by: Sean Paul <seanpaul@chromium.org>

Signed-off-by: Rob Clark <robdclark@gmail.com>
(cherry picked from commit 18a63b3c6f0b6247f93f5eb05353b0486512017e)
(am from git://anongit.freedesktop.org/drm/drm)

BUG=None
TEST=Tested on cheza, boots and display comes up

Signed-off-by: Sean Paul <seanpaul@chromium.org>
Change-Id: Ibecf4f4915ec847517b0b677a9d1ced479e6c11a
Reviewed-on: https://chromium-review.googlesource.com/1439570
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>