FROMGIT: regulator: core: enable power when setting up constraints
When a regulator is marked as always on, it is enabled early on, when
checking and setting up constraints. It makes the assumption that the
bootloader properly initialized the regulator, and just in case enables
the regulator anyway.

Some constraints however currently get missed, such as the soft-start
and ramp-delay. This causes the regulator to be enabled, without the
soft-start and ramp-delay being applied, which in turn can cause
high-currents or other start-up problems.

By moving the always-enabled constraints later in the constraints check,
we can at least ensure all constraints for the regulator are followed.

Signed-off-by: Olliver Schinagl <oliver@schinagl.nl>
Signed-off-by: Priit Laes <plaes@plaes.org>
Signed-off-by: Mark Brown <broonie@kernel.org>
(cherry picked from commit 2bb1666369339f69f227ad060c250afde94d5c69
 git://git.kernel.org/pub/scm/linux/kernel/git/broonie/regulator.git for-next)

BUG=b:119687265
TEST=Manutally test regulator_set_load() cases

Change-Id: I74420f5423ea996117e89077ce6274c6155c56da
Signed-off-by: Douglas Anderson <dianders@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1351430
Reviewed-by: Matthias Kaehlcke <mka@chromium.org>