FROMGIT: clk: qcom: Add a dummy enable function for GX gdsc
Most of the time the CPU should not be touching the GX
domain on the GPU except for a very special use case when
the CPU needs to force the GX headswitch off. Add a
dummy enable function for the GX gdsc to simulate success
so that the pm_runtime reference counting is correct.

Signed-off-by: Jordan Crouse <jcrouse@codeaurora.org>
Signed-off-by: Stephen Boyd <sboyd@kernel.org>
(cherry picked from commit 85a3d920d30a5460522d180205a61aee53687d07
 git://git.kernel.org/pub/scm/linux/kernel/git/clk/linux.git clk-qcom-gpucc)

BUG=b:79946281
TEST=w/ series GPU works

Change-Id: I18457e415288cbc9e87f3af185f66b6d5dde6b57
Signed-off-by: Douglas Anderson <dianders@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1353071
Reviewed-by: Stephen Boyd <swboyd@chromium.org>