FROMGIT: dt-bindings: clock: qcom: Fix the xo parent in gpucc example
In the bindings that landed for the gpucc we require that the XO clock
(one possible parent of the gpucc) be listed.  The code doesn't use
this yet--this is just to allow us to move toward the day when it does
use it.  What the code does do today is to hardcode the parent name to
"bi_tcxo".  That's all well and good.

...but the example in the bindings shows this clock mapping to the
clock "xo_board".  On the current sdm845.dtsi file the "xo_board"
clock is a fixed clock with an output name of "xo_board".  The clock
with the name "bi_tcxo" is actually provided by the RPMh Clock
Controller.  Presumably that's the one that was wanted.

Let's update the example to make this clearer.

Fixes: e431c92188a9 ("dt-bindings: clock: Introduce QCOM Graphics clock bindings")
Signed-off-by: Douglas Anderson <dianders@chromium.org>
Signed-off-by: Stephen Boyd <sboyd@kernel.org>
(cherry picked from commit 922b8fae085d6561ef4a2d1527a14bfde82fd1e8
 git://git.kernel.org/pub/scm/linux/kernel/git/clk/linux.git clk-qcom-gpucc)

BUG=b:79946281
TEST=Example matches DTS

Change-Id: Ib0a2c4eadb655a9c21358751cb2973c30c67a268
Signed-off-by: Douglas Anderson <dianders@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1354086