BACKPORT: FROMLIST: cros_ec: handle MKBP more events flag
CL:1341159 adds a "more events" flag to allow the EC to notify us that
there are more MKBP events waiting to be served. This CL extends the
cros_ec bottom-half IRQ handler to listen to that flag in order to try
and serve all known events before returning to the kernel with IRQ_HANDLED.

This serves an optimization purpose as it minimizes the round-trip back
and forth from the interrupt handler, and it also improves the quality
of sensor timestamping, as it avoids a mismatch of event times between
EC and AP that can confuse the median filter we use in the sensor ring.

As described in the EC CL, this change is forward and backward compatible.

This would be a cherry-pick of CL:1340983, but the code structure is
different enough between kernel versions 4.4 and 4.14 that several
tweaks are required to make things compile and work. It has been sent
upstream as https://lore.kernel.org/patchwork/patch/1019477/

BUG=b:119570064, b:112102925, b:114915792
TEST=verify that sensor events come in and have sensible values and
     timestamps

I had to tweak the merge of cros_ec_proto, but it was straight forward.

Signed-off-by: Enrico Granata <egranata@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1358742
Reviewed-by: Gwendal Grignou <gwendal@chromium.org>

Change-Id: Ie51c5074489d3f076a907da30762e2f604c0e6f1
Signed-off-by: Jett Rink <jettrink@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1551940
Reviewed-by: Enrico Granata <egranata@chromium.org>