UPSTREAM: HID: input: use the Resolution Multiplier for high-resolution scrolling
Windows uses a magic number of 120 for a wheel click. High-resolution
scroll wheels are supposed to use a fraction of 120 to signal smaller
scroll steps. This is implemented by the Resolution Multiplier in the
device itself.

If the multiplier is present in the report descriptor, set it to the
logical max and then use the resolution multiplier to calculate the
high-resolution events. This is the recommendation by Microsoft, see
http://msdn.microsoft.com/en-us/windows/hardware/gg487477.aspx

Note that all mice encountered so far have a logical min/max of 0/1, so
it's a binary "yes or no" to high-res scrolling anyway.

To make userspace simpler, always enable the REL_WHEEL_HI_RES bit. Where
the device doesn't support high-resolution scrolling, the value for the
high-res data will simply be a multiple of 120 every time. For userspace,
if REL_WHEEL_HI_RES is available that is the one to be used.

Potential side-effect: a device with a Resolution Multiplier applying to
other Input items will have those items set to the logical max as well.
This cannot easily be worked around but it is doubtful such devices exist.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>
(am from https://patchwork.kernel.org/patch/10712841/)
(also found at https://lkml.kernel.org/r/20181205004228.10714-5-peter.hutterer@who-t.net)
(cherry picked from commit 2dc702c991e3774af9d7ce410eef410ca9e2357e)

BUG=chromium:888172
TEST=As part of high-resolution scrolling series, checked that a Microsoft
     Comfort Optical Mouse 3000 reported REL_WHEEL_HI_RES events of 30
     each during vertical scrolling. Also checked various Logitech mice
     with the other commits in the series, and checked that low-resolution
     scrolling (vertical or horizontal) reported events of 120 on the
     appropriate _HI_RES axis.

Change-Id: I8c46fa52cb352036d5d5d5617cd56e6f8f9a8669
Signed-off-by: Harry Cutts <hcutts@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1479921
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Dmitry Torokhov <dtor@chromium.org>