UPSTREAM: kdb: Don't back trace on a cpu that didn't round up
If you have a CPU that fails to round up and then run 'btc' you'll end
up crashing in kdb becaue we dereferenced NULL.  Let's add a check.
It's wise to also set the task to NULL when leaving the debugger so
that if we fail to round up on a later entry into the debugger we
won't backtrace a stale task.

Signed-off-by: Douglas Anderson <dianders@chromium.org>
Acked-by: Daniel Thompson <daniel.thompson@linaro.org>
Signed-off-by: Daniel Thompson <daniel.thompson@linaro.org>
(cherry picked from commit 162bc7f5afd75b72acbe3c5f3488ef7e64a3fe36)

BUG=chromium:900598
TEST=Run kgdb w/ lockdep

Change-Id: I15301ed573b6d83a3f7e7c5392bc3689396c43d4
Signed-off-by: Douglas Anderson <dianders@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1399185
Reviewed-by: Guenter Roeck <groeck@chromium.org>