CHROMIUM: iwl7000: mvm: don't hide HE radiotap data in SKB
Hiding the HE radiotap data for further processing of the SKB just
caused another bug when adding the L-SIG data. Simply stop doing
this and adjust the skb->data pointer accordingly when we need to
get the 802.11 header.

While at it, also verify and fix the data alignment, we need to add
2 bytes padding with the vendor data to ensure the whole length of
all radiotap headers is a multiple of 4.

Change-Id: I76218f6d55e464a52c1f5e7469abb82d26d4e471
Signed-off-by: Johannes Berg <johannes.berg@intel.com>
iwl7000-tree: 1516c20f85afa58e1cd7327c0a333842590a99d9