FROMGIT: OPP: Use opp_table->regulators to verify no regulator case
The value of opp_table->regulator_count is not very consistent right now
and it may end up being 0 while we do have a "opp-microvolt" property in
the OPP table. It was kept that way as we used to check if any
regulators are set with the OPP core for a device or not using value of
regulator_count.

Lets use opp_table->regulators for that purpose as the meaning of
regulator_count is going to change in the later patches.

Reported-by: Quentin Perret <quentin.perret@arm.com>
Signed-off-by: Viresh Kumar <viresh.kumar@linaro.org>
(cherry picked from commit 90e3577b5feb42bac1269e16bb3d2bdd8f6df40f
 git://git.kernel.org/pub/scm/linux/kernel/git/vireshk/pm.git opp/linux-next)

BUG=b:80592702
TEST=Build and boot

Change-Id: I851b79635c831db9446ce4e2bddc9df64548d34b
Signed-off-by: Douglas Anderson <dianders@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1384713
Reviewed-by: Stephen Boyd <swboyd@chromium.org>