UPSTREAM: ALSA: hda/hdmi: Always set display_power_control for Intel HSW+ codecs
We've excluded the display_power_control flag for Intel HSW and BDW
codecs as the HD-audio controllers of the corresponding platforms take
care of the display power as well.  But the recent refactoring
separates the controller and the codec power accounting, so it's fine
to call the display PM even for HSW/BDW codecs.  This is less
confusing since we can avoid this well-hidden condition.

Signed-off-by: Takashi Iwai <tiwai@suse.de>
(cherry picked from commit 46594d3345f0432ddc83b9f3c1b04492fdfafd07)

BUG=b:123738217
TEST=Test Audio use cases for CML with full SOF patch series applied.

Change-Id: I8642d2a083c20dfccc62d3332e2ad6c0d0430df6
Signed-off-by: smit shah <smit.shah@intel.com>
Signed-off-by: Ap, Kamal <kamal.ap@intel.corp-partner.google.com>
Signed-off-by: Ben Zhang <benzh@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1499107
Reviewed-by: Sean Paul <seanpaul@chromium.org>