FIXUP: FROMLIST: interconnect: Add generic on-chip interconnect API
===
This FIXUP updates to v12 of the series on the list. It was created by
reverting:
commit da4c1b2d9f5ee ("BACKPORT: FROMLIST: interconnect: Add generic
on-chip interconnect API"),
reapplying via fromupstream.py, and squashing the two together.

===

This patch introduces a new API to get requirements and configure the
interconnect buses across the entire chipset to fit with the current
demand.

The API is using a consumer/provider-based model, where the providers are
the interconnect buses and the consumers could be various drivers.
The consumers request interconnect resources (path) between endpoints and
set the desired constraints on this data flow path. The providers receive
requests from consumers and aggregate these requests for all master-slave
pairs on that path. Then the providers configure each node along the path
to support a bandwidth that satisfies all bandwidth requests that cross
through that node. The topology could be complicated and multi-tiered and
is SoC specific.

Reviewed-by: Evan Green <evgreen@chromium.org>
Signed-off-by: Georgi Djakov <georgi.djakov@linaro.org>
(am from https://lore.kernel.org/patchwork/patch/1022905/)
(also found at https://lkml.kernel.org/r/20181208170216.32555-2-georgi.djakov@linaro.org)

Conflicts:
   drivers/Kconfig
   drivers/Makefile

BUG=b:80591049
TEST=Boot cheza

Change-Id: I0d6ff82db64c4be8366b0e8c73ac7c5b615ef3a8
Signed-off-by: Evan Green <evgreen@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1372581
Commit-Ready: Douglas Anderson <dianders@chromium.org>
Reviewed-by: Douglas Anderson <dianders@chromium.org>