ANDROID: sched: Align EAS with upstream
The core EAS patches have now been accepted upstream. The patches used
in Android are based on a slightly earlier version of the series. In
order to reduce the delta with mainline and ease backports, align the
EAS code paths with their upstream version.

This basically applies the output of git range-diff on the appropriate
commits, and fixes a conflict in schedutil regarding the integration of
schedtune.

Bug: 120440300
Signed-off-by: Quentin Perret <quentin.perret@arm.com>
(cherry-picked from commit 81dd278ee46713591ab5d6866793b5f0c7ffeea4)

BUG=b:122901369, b:111527336
TEST=build for cheza

Change-Id: I208ebeb4207e3f4f4bbb5103c606b293a464c20f
Signed-off-by: Matthias Kaehlcke <mka@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1418083
Reviewed-by: Evan Green <evgreen@chromium.org>