CHROMIUM: arm64: dts: qcom: sdm845-cheza: Enable pulls for BT RX and CTS lines
The Bluetooth module only drives its TX pin when it is (fully)
powered on. Enable a pull-up on the AP side (RX / GPIO 48) to avoid
reading garbage data from a tristate signal.

Also enable a pull-down on the AP CTS pin (GPIO 45) to match the
pull-down on the RTS signal of the Bluetooth controller.

This change partially undoes CL:1312750, which removed pull-ups from
CTS and RX. Key of that CL was the removal of the pull-up from CTS,
which essentially remains in place.

BUG=b:118875972, b:111748385
TEST=with CL:1319450 and CL:1319451
  bring hci0 up/down repeatedly
    => no more "frame reassembly" errors

Change-Id: Ifc1c666760964154a09c93d20b975af42cbd8fdb
Signed-off-by: Matthias Kaehlcke <mka@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1391888
Commit-Ready: Douglas Anderson <dianders@chromium.org>
Reviewed-by: Alexandru M Stan <amstan@chromium.org>
Reviewed-by: Douglas Anderson <dianders@chromium.org>